% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_e0.R
\name{f.e0}
\alias{f.e0}
\title{Compute e0 (life expectancy at age zero) values from lx values of the observed mortality dataframe}
\usage{
f.e0(df, lx, age, g, e0u)
}
\arguments{
\item{df}{data frame containing lx values}

\item{lx}{variable (str) containing lx values in 100,000 people}

\item{age}{variable (str) containing age information}

\item{g}{specify gender (str); "F" or "FR" or "FU" = female (general, rural, urban); "M" or "MR" or "MU" = male (general, rural, urban); only needed for mx max computation}

\item{e0u}{life expectancy at age zero e0 (numeric)}
}
\value{
e0 (life expectancy at age zero)
}
\description{
Compute e0 (life expectancy at age zero) values from lx values of the observed mortality dataframe
}
