% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_in_dom_mig.R
\name{f.in.dom.mig.calc}
\alias{f.in.dom.mig.calc}
\title{Function to calculate the total in migration numbers for all states}
\usage{
f.in.dom.mig.calc(input.path, pop.dataframe, scen.factor)
}
\arguments{
\item{input.path}{input folder containing the migration rates}

\item{pop.dataframe}{population data frame}

\item{scen.factor}{scale applied to migration rates based on the scenario}
}
\value{
data frame containing in-migration values for all states
}
\description{
Function to calculate the total in migration numbers for all states
}
