% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_in_state_dom_mig.R
\name{f.in.state.dom.mig.calc}
\alias{f.in.state.dom.mig.calc}
\title{Function to disaggregate total in-migration to each state}
\usage{
f.in.state.dom.mig.calc(input.path, state, pop.dataframe, scen.factor)
}
\arguments{
\item{input.path}{input folder containing the migration rates}

\item{state}{current state}

\item{pop.dataframe}{population data frame}

\item{scen.factor}{scale applied to migration rates based on the scenario}
}
\value{
data frame containing in-migration values to the current state from all other states
}
\description{
Function to disaggregate total in-migration to each state
}
