% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_lbx.R
\name{f.lBx}
\alias{f.lBx}
\title{Function to compute matrix representation of birth rates (Bx)}
\usage{
f.lBx(dfFert, dfSrR, dfSrU, Lx, lx, SxF, t, g)
}
\arguments{
\item{dfFert}{Data Frame containing fertility variables in the form "tpFx"+region+step (e.g., "tpFxR2")}

\item{dfSrR}{Data frame containing sex ratios for rural area (variables: propM and propF)}

\item{dfSrU}{Data frame containing sex ratios for urban area (variables: propM and propF)}

\item{Lx}{nested list containing maxtrices of Lx values}

\item{lx}{nested list containing maxtrices of lx values}

\item{SxF}{nested list containing maxtrices of Sx values for females}

\item{t}{time step used in the projection}

\item{g}{gender "F"=female; "M"=male}
}
\value{
matrix representation of birth rates (Bx)
}
\description{
Function to compute matrix representation of birth rates (Bx)
}
