% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_lin_int.R
\name{f.linIntE}
\alias{f.linIntE}
\title{Function to perform linear interpolation of rates/values. Uses 5-year intervals to compute values for 1-year intervals}
\usage{
f.linIntE(x, valN, si)
}
\arguments{
\item{x}{vector of values}

\item{valN}{name of variable (str) to hold 1-year values}

\item{si}{should the first value be used for the start of the first interval (e.g., year=0) -> TRUE; or at the center of the first interval (e.g., year=2) -> FALSE}
}
\value{
linearly interpolated values at 1-year interval
}
\description{
Function to perform linear interpolation of rates/values. Uses 5-year intervals to compute values for 1-year intervals
}
