% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_lint.R
\name{f.lInt}
\alias{f.lInt}
\title{Linear interpolation}
\usage{
f.lInt(df, x, y, xval)
}
\arguments{
\item{df}{data frame}

\item{x}{variable in df that contains val}

\item{y}{variable in df for which value should be interpolated}

\item{xval}{value on x for which a corresponding value of y should be found}
}
\value{
y value that corresponds to xval
}
\description{
Linear interpolation
}
