% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_lin_columns.R
\name{f.lIntC}
\alias{f.lIntC}
\title{Linear interpolation of columns}
\usage{
f.lIntC(df, x1, x2, y1, y2, yi)
}
\arguments{
\item{df}{data frame containing x1 and x2}

\item{x1}{variable containing the x values at level y1}

\item{x2}{variable containing the x values at level y2}

\item{y1}{level of y for x1 (number)}

\item{y2}{level of y for x2 (number)}

\item{yi}{level of y for which x should be interpolated (number)}
}
\value{
vector containing the imputed values of xi
}
\description{
Linear interpolation of columns
}
