% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ltm.R
\name{f.ltm}
\alias{f.ltm}
\title{Function to compute mortality variables as matrices}
\usage{
f.ltm(dfMort, dfDomMig, t, g)
}
\arguments{
\item{dfMort}{Data Frame containing mortality variables in the form "mx"+gender+region+year (e.g., "mxFR2")}

\item{dfDomMig}{Data Frame containing variables on domestic (internal) migration propabilities in the form "dm"+gender+region+year (e.g., "dmMUR0")}

\item{t}{year for which mortality data is needed in the projection}

\item{g}{gender "F"=female; "M"=male}
}
\value{
mortality variables
}
\description{
Function to compute mortality variables as matrices
}
