% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_mov_avg.R
\name{f.movAv}
\alias{f.movAv}
\title{Moving average. Assumes that data frame is sorted correctly}
\usage{
f.movAv(df, var, vart, ws = 3)
}
\arguments{
\item{df}{data frame}

\item{var}{variable for which moving average should be computed}

\item{vart}{time variable that uniquely identifies a time step within each group (same length)}

\item{ws}{window size (default=3); must be odd number (integer)}
}
\value{
smoothed data frame
}
\description{
Moving average. Assumes that data frame is sorted correctly
}
