% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_mx_max.R
\name{f.mxMax}
\alias{f.mxMax}
\title{Compute mx value of last age category (e.g., 100+) based on user specified life expectancy at age 0 for the base year}
\usage{
f.mxMax(g, e0u)
}
\arguments{
\item{g}{gender (str); "F" or "FR" or "FU" = female (general, rural, urban); "M" or "MR" or "MU" = male (general, rural, urban)}

\item{e0u}{life expectancy at age zero (numeric)}
}
\value{
mx value of last age category
}
\description{
Compute mx value of last age category (e.g., 100+) based on user specified life expectancy at age 0 for the base year
}
