% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_out_dom_mig.R
\name{f.out.dom.mig.calc}
\alias{f.out.dom.mig.calc}
\title{Function to calculate the total out migration numbers for all states}
\usage{
f.out.dom.mig.calc(input.path, pop.dataframe, scen.factor)
}
\arguments{
\item{input.path}{input folder containing the migration rates}

\item{pop.dataframe}{population data frame}

\item{scen.factor}{scale applied to migration rates based on the scenario}
}
\value{
data frame containing out-migration values for all states
}
\description{
Function to calculate the total out migration numbers for all states
}
