% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_rep_id.R
\name{f.repID}
\alias{f.repID}
\title{Replace IDs}
\usage{
f.repID(dfLabs, dfRep, varO, varN, varR)
}
\arguments{
\item{dfLabs}{data frame containing labels}

\item{dfRep}{data frame containing the variable to be relabeled}

\item{varO}{variable in dfLabs containing the original ID}

\item{varN}{variable in dfLabs containing the new ID}

\item{varR}{variable in dfRep containing the original ID that should be replaced with the new ID}
}
\value{
The replaced data frame
}
\description{
Replace IDs
}
